requirements.txt
fastapi
uvicorn
pandas
numpy
pydantic
httpx
pyyaml
